Initialize all non-list properties for new instances.
Fixes problem where a rule would not run on new instances in some cases.  At one point in the distant past, rules did not run when they were first created, only when their inputs fired events.  This could be problematic since when a property was added and calculated it might be eligible for calculation immediately for existing instances.  However, the property events for the inputs to the calculated rule would not fire (since they had fired previously), and so the rule would not run.  In order to get around this only server-origin properties were inited when objects were created.  This allowed a calculation to run on property get since the property had not been inited when accessed.  This worked under the assumption that these calculations were being registered for client-origin properties.  This initialization approach is no longer necessary since calculations are performed when created.