Address memory leaks
Expand on cleanup/disposal of components and controls to avoid potential memory leaks in IE.  Swap order of ifNull and format methods so that null values can leverage the null value option passed into the markup extension.  Clean up binding options passed into "#" markup extension and do not pass them into the adapter that is created.  Register the adapter as a disposable component when the "#" markup extension is used.