Fix for loading error that could occur in specific cases.
Back-to-back list add/remove change events may not apply in the same order. This was caused by a signal.pending call within applyListChange for list adds. Since the signal callback could be async, in that case a following list remove might actually be performed prior to the list add, which would result in incorrect list state. It appears that the argument to signal pending to have it execute sync was passed to the wrong function, so moved the argument to the correct function call.